# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClipFlow API",
    "description": "A FastAPI backend service for video editing that analyzes audio tracks to detect speech segments, silence, and optionally filler words.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/analysis/media/{media_id}/analyze": {
      "post": {
        "tags": ["Analysis"],
        "summary": "Start Analysis",
        "description": "Start audio analysis for a media file.",
        "operationId": "start_analysis_api_v1_analysis_media__media_id__analyze_post",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AnalysisCreate" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Analysis Started",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AnalysisStartResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "AI Features Disabled"
          },
          "404": {
            "description": "Media Not Found"
          }
        }
      }
    },
    "/api/v1/analysis/{analysis_id}": {
      "get": {
        "tags": ["Analysis"],
        "summary": "Get Analysis",
        "description": "Get analysis result.",
        "operationId": "get_analysis_api_v1_analysis__analysis_id__get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AnalysisResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/analysis/{analysis_id}/status": {
      "get": {
        "tags": ["Analysis"],
        "summary": "Get Analysis Status",
        "description": "Get lightweight analysis status for polling.",
        "operationId": "get_analysis_status_api_v1_analysis__analysis_id__status_get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AnalysisStatusResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProcessingMode": {
        "type": "string",
        "enum": ["vad", "whisper"]
      },
      "AnalysisStatus": {
        "type": "string",
        "enum": ["pending", "processing", "completed", "failed"]
      },
      "AnalysisCreate": {
        "type": "object",
        "properties": {
          "processing_mode": {
            "type": "string",
            "enum": ["vad", "whisper", "auto"],
            "default": "auto"
          },
          "vad_aggressiveness": { "type": "integer", "minimum": 1, "maximum": 3, "nullable": true },
          "min_silence_duration_ms": { "type": "integer", "minimum": 100, "maximum": 5000, "nullable": true },
          "min_speech_duration_ms": { "type": "integer", "minimum": 100, "maximum": 5000, "nullable": true },
          "detect_filler_words": { "type": "boolean", "nullable": true },
          "custom_filler_words": {
            "type": "array",
            "items": { "type": "string" },
            "nullable": true
          }
        }
      },
      "AnalysisStartResponse": {
        "type": "object",
        "properties": {
          "analysis_id": { "type": "string", "format": "uuid" },
          "status": { "$ref": "#/components/schemas/AnalysisStatus" }
        }
      },
      "SegmentResponse": {
        "type": "object",
        "properties": {
          "start_ms": { "type": "integer" },
          "end_ms": { "type": "integer" },
          "type": { "type": "string", "enum": ["speech", "silence", "filler"] },
          "confidence": { "type": "number" },
          "text": { "type": "string", "nullable": true },
          "filler_word": { "type": "string", "nullable": true }
        }
      },
      "FillerWordDetection": {
        "type": "object",
        "properties": {
          "word": { "type": "string" },
          "start_ms": { "type": "integer" },
          "end_ms": { "type": "integer" },
          "confidence": { "type": "number" }
        }
      },
      "AnalysisResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "media_file_id": { "type": "string", "format": "uuid" },
          "processing_mode": { "$ref": "#/components/schemas/ProcessingMode" },
          "status": { "$ref": "#/components/schemas/AnalysisStatus" },
          "segments": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SegmentResponse" },
            "nullable": true
          },
          "transcription": { "type": "string", "nullable": true },
          "filler_words_detected": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FillerWordDetection" },
            "nullable": true
          },
          "processing_time_ms": { "type": "integer", "nullable": true },
          "error_message": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" },
          "completed_at": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "AnalysisStatusResponse": {
        "type": "object",
        "properties": {
          "status": { "$ref": "#/components/schemas/AnalysisStatus" },
          "progress_percent": { "type": "integer", "nullable": true },
          "error_message": { "type": "string", "nullable": true }
        }
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}